move further funs to util
